+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Please refer to [UPGRADING](UPGRADING.md) guide for upgrading to a major version
 
 ## 7.11.0 - Upcoming
 
+### Added
+
+- Added the `protocols` request option to restrict URI schemes for request transfers
+
 ### Changed
 
 - Adjusted `guzzlehttp/psr7` version constraint to `^2.10`